Fixed update scripts for STRING and Taxonomy.

+++ server_update/updateSTRING.py
@@ -1,7 +1,8 @@
 ##!interval=7
 ##!contact=ales.erjavec@fri.uni-lj.si
 
-import obiPPI, orngServerFiles
+from Orange.bio import obiPPI
+import Orange.utils.serverfiles as orngServerFiles
 import os, sys, shutil, urllib2, gzip
 from getopt import getopt
 
@@ -12,24 +13,23 @@ password = opt.get("-p", opt.get("--password", "password"))
 
 serverFiles = orngServerFiles.ServerFiles(username, password)
 
-import obiPPI
-
 filename = orngServerFiles.localpath("PPI", obiPPI.STRING.FILENAME)
 
-if os.path.exists(filename):
-    os.remove(filename)
+if False:
+    if os.path.exists(filename):
+        os.remove(filename)
 
-obiPPI.STRING.download_data("v9.0")
+    obiPPI.STRING.download_data("v9.0")
 
-gzfile = gzip.GzipFile(filename + ".gz", "wb")
-shutil.copyfileobj(open(filename, "rb"), gzfile)
+    gzfile = gzip.GzipFile(filename + ".gz", "wb")
+    shutil.copyfileobj(open(filename, "rb"), gzfile)
 
-serverFiles.upload("PPI", obiPPI.STRING.FILENAME, filename + ".gz", 
-                   "STRING Protein interactions (Creative Commons Attribution 3.0 License)",
-                   tags=["protein interaction", "STRING", 
-                         "#compression:gz", "#version:%s" % obiPPI.STRING.VERSION]
-                   )
-serverFiles.unprotect("PPI", obiPPI.STRING.FILENAME)
+    serverFiles.upload("PPI", obiPPI.STRING.FILENAME, filename + ".gz", 
+                       "STRING Protein interactions (Creative Commons Attribution 3.0 License)",
+                       tags=["protein interaction", "STRING", 
+                             "#compression:gz", "#version:%s" % obiPPI.STRING.VERSION]
+                       )
+    serverFiles.unprotect("PPI", obiPPI.STRING.FILENAME)
 
 # The second part
 filename = orngServerFiles.localpath("PPI", obiPPI.STRINGDetailed.FILENAME_DETAILED)
